Remove ActionFilterPath
