--- go/lib/pktcls/testdata/act_1.json
@@ -1,11 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondAllOf": [
-                {
-                    "CondBool": true
-                }
-            ]
-        }
-    }
-}
