--- go/lib/pktcls/testdata/act_10.json
@@ -1,1 +1,0 @@
-null
