--- go/lib/pktcls/testdata/act_2.json
@@ -1,15 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondAllOf": [
-                {
-                    "CondAllOf": [
-                        {
-                            "CondBool": true
-                        }
-                    ]
-                }
-            ]
-        }
-    }
-}
