--- go/lib/pktcls/testdata/act_3.json
@@ -1,15 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondAnyOf": [
-                {
-                    "CondAnyOf": [
-                        {
-                            "CondBool": false
-                        }
-                    ]
-                }
-            ]
-        }
-    }
-}
