--- go/lib/pktcls/testdata/act_4.json
@@ -1,9 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondNot": {
-                "CondBool": false
-            }
-        }
-    }
-}
