--- go/lib/pktcls/testdata/act_5.json
@@ -1,16 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondAnyOf": [
-                {
-                    "CondBool": false
-                },
-                {
-                    "CondPathPredicate": {
-                        "PP": "1-1#123"
-                    }
-                }
-            ]
-        }
-    }
-}
