--- go/lib/pktcls/testdata/act_6.json
@@ -1,20 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondAnyOf": [
-                {
-                    "CondBool": false
-                }
-            ]
-        }
-    },
-    "B": {
-        "ActionFilterPaths": {
-            "CondAllOf": [
-                {
-                    "CondBool": true
-                }
-            ]
-        }
-    }
-}
