--- go/lib/pktcls/testdata/act_8.json
@@ -1,18 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondAnyOf": [
-                {
-                    "CondPathPredicate": {
-                        "PP": "1-2#123"
-                    }
-                },
-                {
-                    "CondPathPredicate": {
-                        "PP": "2-1#134"
-                    }
-                }
-            ]
-        }
-    }
-}
