--- go/lib/pktcls/testdata/act_9.json
@@ -1,36 +1,0 @@
-{
-    "A": {
-        "ActionFilterPaths": {
-            "CondAnyOf": [
-                {
-                    "CondAllOf": [
-                        {
-                            "CondPathPredicate": {
-                                "PP": "1-2#123"
-                            }
-                        },
-                        {
-                            "CondPathPredicate": {
-                                "PP": "2-1#234"
-                            }
-                        }
-                    ]
-                },
-                {
-                    "CondAllOf": [
-                        {
-                            "CondPathPredicate": {
-                                "PP": "1-2#345"
-                            }
-                        },
-                        {
-                            "CondPathPredicate": {
-                                "PP": "2-1#456"
-                            }
-                        }
-                    ]
-                }
-            ]
-        }
-    }
-}
